refactor(agents): extract AgentFilesDelegate + add AgentEditorStateHandle
Introduce AgentEditorStateHandle (mirroring ChatStateHandle) as the shared
state accessor for the editor's delegates, and carve the file concerns into
AgentFilesDelegate: per-capability attachment slots (upload/remove with
tool_resource routing), avatar upload/reset, and the lazy file-metadata
enrichment cache. The loader's post-load re-merge and the revert reload now
call the delegate (remergeLoadedFiles / resetFileCache).

No behavior change.
