+++ feature/agents/src/commonMain/kotlin/com/garfiec/librechat/feature/agents/viewmodel/AgentEditorViewModel.kt
@@ -18,7 +18,6 @@ import com.garfiec.librechat.core.model.Agent
 import com.garfiec.librechat.core.model.AgentCategory
 import com.garfiec.librechat.core.model.AgentFile
 import com.garfiec.librechat.core.model.AgentSubagentsConfig
-import com.garfiec.librechat.core.model.FileObject
 import com.garfiec.librechat.core.model.HandoffEdge
 import com.garfiec.librechat.core.model.SkillSummary
 import com.garfiec.librechat.core.model.SupportContact
@@ -28,7 +27,6 @@ import com.garfiec.librechat.core.model.permissions.PermissionType
 import com.garfiec.librechat.core.model.permissions.hasAccessOrPermissive
 import com.garfiec.librechat.core.model.request.CreateActionRequest
 import com.garfiec.librechat.core.model.request.CreateAgentRequest
-import com.garfiec.librechat.core.model.request.DeleteFileEntry
 import com.garfiec.librechat.core.model.request.FunctionTool
 import com.garfiec.librechat.core.model.request.RevertAgentRequest
 import com.garfiec.librechat.core.model.request.UpdateAgentRequest
@@ -43,7 +41,7 @@ import com.garfiec.librechat.feature.agents.components.model.AgentVersion
 import com.garfiec.librechat.feature.agents.components.model.AgentVisibility
 import com.garfiec.librechat.feature.agents.components.model.SupportContactState
 import com.garfiec.librechat.feature.agents.util.ContentReader
-import kotlinx.coroutines.CancellationException
+import com.garfiec.librechat.feature.agents.viewmodel.delegate.AgentFilesDelegate
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -53,7 +51,6 @@ import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 
@@ -246,12 +243,15 @@ class AgentEditorViewModel(
     private val _events = MutableSharedFlow<AgentEditorEvent>()
     val events: SharedFlow<AgentEditorEvent> = _events.asSharedFlow()
 
-    /** Cached file-detail payload from `GET /api/files/agent/:id`. Held here so
-     *  that whichever of [loadAgent] / [loadAgentFiles] finishes second can
-     *  trigger the merge against the by-then-populated tool_resources slots.
-     *  Without this cache, [loadAgentFiles] finishing first would read empty
-     *  slot lists and produce a no-op enrichment. */
-    private var loadedAgentFileObjects: List<FileObject>? = null
+    private val stateHandle = AgentEditorStateHandle(_uiState, viewModelScope)
+
+    private val filesDelegate = AgentFilesDelegate(
+        stateHandle = stateHandle,
+        agentRepository = agentRepository,
+        fileRepository = fileRepository,
+        contentReader = contentReader,
+        ioDispatcher = ioDispatcher,
+    )
 
     init {
         loadAvailableTools()
@@ -268,45 +268,8 @@ class AgentEditorViewModel(
         if (editAgentId != null) {
             loadAgent(editAgentId)
             loadActions()
-            loadAgentFiles(editAgentId)
+            filesDelegate.loadAgentFiles(editAgentId)
         }
-    }
-
-    /**
-     * Fetches `GET /api/files/agent/:id` and merges filename/bytes/type into
-     * the per-capability slots populated from `tool_resources.<X>.file_ids`.
-     * The agent payload only carries file_ids — without this call, chips
-     * would render with bare IDs.
-     */
-    private fun loadAgentFiles(agentId: String) {
-        viewModelScope.launch {
-            when (val result = fileRepository.getAgentFiles(agentId)) {
-                is Result.Success -> {
-                    loadedAgentFileObjects = result.data
-                    mergeAgentFileMetadata(result.data)
-                }
-                is Result.Error -> { /* Best-effort enrichment; chips fall back to fileId */ }
-                is Result.Loading -> { /* no-op */ }
-            }
-        }
-    }
-
-    private fun mergeAgentFileMetadata(files: List<FileObject>) {
-        val byId = files.associateBy { it.fileId }
-        fun enrich(list: List<AgentFile>): List<AgentFile> = list.map { agentFile ->
-            byId[agentFile.fileId]?.let { obj ->
-                agentFile.copy(
-                    filename = agentFile.filename ?: obj.filename,
-                    bytes = agentFile.bytes ?: obj.bytes,
-                    type = agentFile.type ?: obj.type,
-                )
-            } ?: agentFile
-        }
-        _uiState.value = _uiState.value.copy(
-            codeFiles = enrich(_uiState.value.codeFiles),
-            knowledgeFiles = enrich(_uiState.value.knowledgeFiles),
-            contextFiles = enrich(_uiState.value.contextFiles),
-        )
     }
 
     /**
@@ -539,7 +502,7 @@ class AgentEditorViewModel(
                     // the per-capability slot lists are populated. Without
                     // this, an earlier-finishing files request would have
                     // merged against empty lists and produced no enrichment.
-                    loadedAgentFileObjects?.let { mergeAgentFileMetadata(it) }
+                    filesDelegate.remergeLoadedFiles()
                     Logger.d {
                         "AgentEditor: UI state AFTER mapping - " +
                             "name=${newState.name}, description=${newState.description}, " +
@@ -930,132 +893,9 @@ class AgentEditorViewModel(
 
     // --- Per-capability file attachments ---
 
-    /**
-     * Picks a slot's current list. Used by upload/remove so the caller doesn't
-     * have to switch on the enum.
-     */
-    private fun filesFor(slot: AgentFileSlot): List<AgentFile> = when (slot) {
-        AgentFileSlot.CODE -> _uiState.value.codeFiles
-        AgentFileSlot.KNOWLEDGE -> _uiState.value.knowledgeFiles
-        AgentFileSlot.CONTEXT -> _uiState.value.contextFiles
-    }
+    fun uploadAgentFile(fileRef: Any, slot: AgentFileSlot) = filesDelegate.uploadAgentFile(fileRef, slot)
 
-    private fun setFilesFor(slot: AgentFileSlot, files: List<AgentFile>) {
-        _uiState.value = when (slot) {
-            AgentFileSlot.CODE -> _uiState.value.copy(codeFiles = files)
-            AgentFileSlot.KNOWLEDGE -> _uiState.value.copy(knowledgeFiles = files)
-            AgentFileSlot.CONTEXT -> _uiState.value.copy(contextFiles = files)
-        }
-    }
-
-    /**
-     * Upload a file for the given capability slot. The backend attaches the
-     * file to `tool_resources.<wire>.file_ids` on the agent when both
-     * `agent_id` + `tool_resource` are supplied. New (unsaved) agents can't
-     * accept files yet — the user is told to save first via a snackbar.
-     */
-    fun uploadAgentFile(fileRef: Any, slot: AgentFileSlot) {
-        val agentId = _uiState.value.agentId
-        if (agentId.isNullOrBlank()) {
-            _uiState.value = _uiState.value.copy(
-                error = AGENT_FILES_SAVE_FIRST_MARKER,
-            )
-            return
-        }
-        viewModelScope.launch {
-            try {
-                val bytes = contentReader.readBytes(fileRef) ?: run {
-                    _uiState.value = _uiState.value.copy(error = "Could not read file")
-                    return@launch
-                }
-                if (bytes.size > AGENT_FILE_SIZE_LIMIT_BYTES) {
-                    val limitMb = (AGENT_FILE_SIZE_LIMIT_BYTES / (1024 * 1024)).toInt()
-                    _uiState.value = _uiState.value.copy(
-                        error = "$AGENT_FILES_TOO_LARGE_MARKER$limitMb",
-                    )
-                    return@launch
-                }
-                val filename = contentReader.getFileName(fileRef) ?: "upload"
-                val mimeType = contentReader.getMimeType(fileRef) ?: "application/octet-stream"
-
-                _uiState.value = _uiState.value.copy(
-                    uploadingSlots = _uiState.value.uploadingSlots + slot,
-                )
-                val result = fileRepository.uploadFile(
-                    bytes = bytes,
-                    filename = filename,
-                    type = mimeType,
-                    endpoint = "agents",
-                    agentId = agentId,
-                    toolResource = slot.wire,
-                )
-                _uiState.value = _uiState.value.copy(
-                    uploadingSlots = _uiState.value.uploadingSlots - slot,
-                )
-                when (result) {
-                    is Result.Success -> {
-                        val obj = result.data
-                        val agentFile = AgentFile(
-                            fileId = obj.fileId,
-                            filename = obj.filename,
-                            bytes = obj.bytes,
-                            type = obj.type,
-                            originResource = slot.wire,
-                        )
-                        // Re-pick latest because the snackbar might have cleared
-                        // state between the upload start and completion.
-                        setFilesFor(slot, filesFor(slot) + agentFile)
-                    }
-                    is Result.Error -> {
-                        _uiState.value = _uiState.value.copy(
-                            error = result.message ?: AGENT_FILE_UPLOAD_FAILED_MARKER,
-                        )
-                    }
-                    is Result.Loading -> { /* no-op */ }
-                }
-            } catch (e: Exception) {
-                Logger.e(e) { "uploadAgentFile: unexpected error" }
-                _uiState.value = _uiState.value.copy(
-                    uploadingSlots = _uiState.value.uploadingSlots - slot,
-                    error = AGENT_FILE_UPLOAD_FAILED_MARKER,
-                )
-            }
-        }
-    }
-
-    fun removeAgentFile(fileId: String, slot: AgentFileSlot) {
-        val agentId = _uiState.value.agentId
-        // Local-only state when there's no agentId yet (we never let this happen
-        // through uploadAgentFile, but stay defensive).
-        if (agentId.isNullOrBlank()) {
-            setFilesFor(slot, filesFor(slot).filterNot { it.fileId == fileId })
-            return
-        }
-        // Optimistically remove; rollback on error.
-        val before = filesFor(slot)
-        setFilesFor(slot, before.filterNot { it.fileId == fileId })
-        viewModelScope.launch {
-            val target = before.firstOrNull { it.fileId == fileId } ?: return@launch
-            // The Context slot UI shows files from both `context` and `ocr`
-            // tool_resources merged. Route the deletion to the slot the file
-            // was actually loaded from so the backend can find and remove
-            // the file_id; falling back to slot.wire for files uploaded in
-            // this session (which the picker writes under slot.wire).
-            val toolResource = target.originResource ?: slot.wire
-            val result = fileRepository.deleteFiles(
-                files = listOf(DeleteFileEntry(fileId = target.fileId, filepath = "")),
-                agentId = agentId,
-                toolResource = toolResource,
-            )
-            if (result is Result.Error) {
-                // Rollback
-                setFilesFor(slot, before)
-                _uiState.value = _uiState.value.copy(
-                    error = result.message ?: AGENT_FILE_REMOVE_FAILED_MARKER,
-                )
-            }
-        }
-    }
+    fun removeAgentFile(fileId: String, slot: AgentFileSlot) = filesDelegate.removeAgentFile(fileId, slot)
 
     // --- Sharing ---
 
@@ -1135,63 +975,9 @@ class AgentEditorViewModel(
 
     // --- Avatar ---
 
-    fun uploadAvatar(uri: Any) {
-        val agentId = _uiState.value.agentId ?: return
-        viewModelScope.launch {
-            try {
-                // Reading bytes off the URI is blocking I/O — keep it off the Main
-                // dispatcher (viewModelScope = Main.immediate) to avoid an ANR on
-                // large images. Mirrors the FileAttachmentDelegate fix.
-                val bytes = withContext(ioDispatcher) { contentReader.readBytes(uri) } ?: return@launch
-                if (bytes.size > AVATAR_SIZE_LIMIT_BYTES) {
-                    val limitMb = AVATAR_SIZE_LIMIT_BYTES / (1024 * 1024)
-                    _uiState.value = _uiState.value.copy(
-                        error = "Avatar must be ${limitMb}MB or smaller",
-                    )
-                    return@launch
-                }
-                val mimeType = contentReader.getMimeType(uri) ?: "image/png"
+    fun uploadAvatar(uri: Any) = filesDelegate.uploadAvatar(uri)
 
-                when (val result = agentRepository.uploadAgentAvatar(agentId, bytes, mimeType)) {
-                    is Result.Success -> {
-                        _uiState.value = _uiState.value.copy(
-                            avatarUrl = result.data.avatarUrl,
-                        )
-                    }
-                    is Result.Error -> {
-                        _uiState.value = _uiState.value.copy(
-                            error = result.message ?: "Failed to upload avatar",
-                        )
-                    }
-                    is Result.Loading -> { /* no-op */ }
-                }
-            } catch (e: CancellationException) {
-                // Cooperative cancellation must propagate (SKIE/iOS requirement).
-                throw e
-            } catch (e: Exception) {
-                _uiState.value = _uiState.value.copy(
-                    error = "Failed to read image: ${e.message}",
-                )
-            }
-        }
-    }
-
-    fun resetAvatar() {
-        val agentId = _uiState.value.agentId ?: return
-        viewModelScope.launch {
-            when (val result = agentRepository.resetAgentAvatar(agentId)) {
-                is Result.Success -> {
-                    _uiState.value = _uiState.value.copy(avatarUrl = null)
-                }
-                is Result.Error -> {
-                    _uiState.value = _uiState.value.copy(
-                        error = result.message ?: "Failed to reset avatar",
-                    )
-                }
-                is Result.Loading -> { /* no-op */ }
-            }
-        }
-    }
+    fun resetAvatar() = filesDelegate.resetAvatar()
 
     // --- Duplicate / Delete / Revert ---
 
@@ -1255,8 +1041,8 @@ class AgentEditorViewModel(
                     // stale enrichment cache and re-fetch /api/files/agent/:id so
                     // the new file_ids resolve to filename/bytes/type instead of
                     // showing bare IDs in the chips.
-                    loadedAgentFileObjects = null
-                    loadAgentFiles(agentId)
+                    filesDelegate.resetFileCache()
+                    filesDelegate.loadAgentFiles(agentId)
                 }
                 is Result.Error -> {
                     _uiState.value = _uiState.value.copy(
